Splidet hook and component for ReferenceRadioButton
